Initial Commit
This commit contains the basic directory structure, as well as a
working "build.hxml" and "src/Main.hx"

At the moment this does nothing more than a "hello world", it simply
prints a line to the output.

The Directory structure we've created:
/src/
/src/model/
/src/controller/
/src/Main.hx
/doc/
/doc/template.xml
/out/
/out/view/
/build.hxml

We have not run the initial "build command" yet.